[FLINK-11952][1/3] Make ChildFirstClassLoader a top-level class in flink-core
